==> 1_aledb_mutation_df.ipynb <==
%matplotlib inline
import matplotlib.pyplot as plt
import pandas as pd
from IPython.display import display
import os, sys, itertools, csv
module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)
from mutil.alemutdf import get_all_sample_mut_df, get_multi_exp_max_freq_mut_df
from mutil.mut import is_coding_mut
from mutil.metadata import get_condition_val_dict, get_condition_field_val_set
from mutil.genome import get_K12_pos_from_BOP27

muts_df = get_all_sample_mut_df("./data/muts/pub_growth/")
# unpublished_muts_df = get_all_sample_mut_df("./data/muts/unpublished/")  # I've currently placed GYD into the public mutation set.
# muts_df = muts_df.append(unpublished_muts_df)
display(set(muts_df["exp"]))

Getting number of samples from raw dataset, since by the end of the NB, the samples are flattened into ALEs by the "ALE-unique mutations" logic

def get_exp_sample_name(exp_name, ale, flask, isolate, tech_rep):
    sample_name = exp_name + " " + str(int(ale)) + " " + str(int(flask)) + " " + str(int(isolate)) + " " + str(int(tech_rep))
    return sample_name

muts_df["sample"] = muts_df.apply(lambda r: get_exp_sample_name(r.exp, r.ale, r.flask, r.isolate, r.tech_rep), axis=1)

len(muts_df["sample"].unique())

# Percentage of samples that are populations
muts_df["sample type"] = muts_df.apply(lambda r: "population" if r["isolate"] == 0 else "clone",axis=1)
df = muts_df[["sample", "sample type"]].copy().drop_duplicates()
len(df[df["sample type"]=="population"])/len(df)

len(muts_df)

# Different experiments have different strings for position (some with commas, some without), therefore going ahead and changing them all to integers
muts_df.Position = muts_df.Position.apply(lambda x: int(str(x).replace(",","")))
muts_df.Position = muts_df.Position.astype(int)

# ensure all mutations are using NC_000913 positions.
genome_pos_transform_exp_l = [
    "ERS",
]
muts_df.Position = muts_df.apply(lambda row: get_K12_pos_from_BOP27(row.Position) if row.exp in genome_pos_transform_exp_l else row.Position, axis=1)

# removing unnecessary column
muts_df = muts_df.drop(columns=["Mut ID"])

muts_df.to_pickle("./data/1_df.pkl")

len(muts_df)
