==> 2_5_annot_pathways.ipynb <==
import pandas as pd

all_muts_df = pd.read_pickle("./data/2_4_df.pkl")
display(all_muts_df.shape, all_muts_df.head())

TU_objects_df = pd.read_pickle("./data/TU_objects_df.pkl")
gene_synonym_df = pd.read_pickle("./data/gene_synonym_df.pkl")
gene_pathway_df = pd.read_pickle("./data/gene_pathway_df.pkl")

from mutil.gene import get_gene_bnum


def _get_pathways(gene_id):
    annots = []
    bnum = get_gene_bnum(gene_id, gene_synonym_df)
    df = gene_pathway_df[gene_pathway_df["refseq_locus_tag"]==bnum]
    for pathway in df["pathway_name"]:
        d = {"name": pathway}
        annots.append(d)
    return annots


def get_pathways(mut_row):
    annots = []
    for feat_d in mut_row["genomic features"]:
        # !!! This part only handles mutations that land within a genomic feature.
        # Need to also implement logic for handling mutations that land within intergenic regions.
        if feat_d["genetic"]:
            annots += _get_pathways(feat_d["RegulonDB ID"])
        # This will catch whether no features exist in the feat_ID_set
        # or if those features are just integenic regions.
        # Get operon genetic feature link dict through overlapping mutation and operon ranges.
        # All genetic features are annotated within both the "genetic features" and "genetic feature links" column.
        # If no integenic genomic feature is hit by a mutation, use the genetic feature range in an overlap analysis with operons.
        else:
            TU_IDs = set(TU_objects_df[TU_objects_df["TU_OBJECT_ID"]==feat_d["RegulonDB ID"]]["TRANSCRIPTION_UNIT_ID"])
            for TU_ID in TU_IDs:
                subset_TU_genes_df = TU_objects_df[(TU_objects_df["TRANSCRIPTION_UNIT_ID"]==TU_ID) & (TU_objects_df["TU_OBJECT_CLASS"]=='GN')]
                for gene_id in subset_TU_genes_df["TU_OBJECT_ID"].drop_duplicates():
                    if str(gene_id) != 'nan' and gene_id != '':  # RegulonDB can return 'nan' or '' gene names
                        annots += _get_pathways(gene_id)
    return annots

all_muts_df["pathways"] = all_muts_df.apply(get_pathways, axis=1)
all_muts_df.head()

TU_df = pd.read_pickle("./data/TU_df.pkl")
operon_df = pd.read_pickle("./data/operon_df.pkl")
operon_df["range"] = operon_df.apply(lambda row: (row["REGULATIONPOSLEFT"], row["REGULATIONPOSRIGHT"]), axis=1)

from mutil.operon import get_operon_ID_set


def _add_links_to_d(bnum, d, op_ID_set):
    for pathway in gene_pathway_df[gene_pathway_df["refseq_locus_tag"]==bnum]["pathway_name"]:
        if pathway not in d.keys(): d[pathway] = []
        d[pathway] += list(op_ID_set)


def get_pathway_operon_links_d(mut_row):
    pathway_operon_d = dict()
    for feat_d in mut_row["genomic features"]:
        if feat_d["genetic"]:
            op_ID_set = get_operon_ID_set(feat_d["RegulonDB ID"], TU_objects_df, TU_df, operon_df)
            if str(feat_d["RegulonDB ID"]) != 'nan' and feat_d["RegulonDB ID"] != '':
                bnum = get_gene_bnum(feat_d["RegulonDB ID"], gene_synonym_df)
                _add_links_to_d(bnum, pathway_operon_d, op_ID_set)
        else:
            TU_IDs = set(TU_objects_df[TU_objects_df["TU_OBJECT_ID"]==feat_d["RegulonDB ID"]]["TRANSCRIPTION_UNIT_ID"])
            for TU_ID in TU_IDs:
                subset_TU_genes_df = TU_objects_df[(TU_objects_df["TRANSCRIPTION_UNIT_ID"]==TU_ID) & (TU_objects_df["TU_OBJECT_CLASS"]=='GN')]

                op_ID_set = set()
                for _, TU_gene in subset_TU_genes_df.iterrows():
                     op_ID_set = op_ID_set | get_operon_ID_set(TU_gene["TU_OBJECT_ID"], TU_objects_df, TU_df, operon_df)

                for gene_id in subset_TU_genes_df["TU_OBJECT_ID"].drop_duplicates():
                    if str(gene_id) != 'nan' and gene_id != '':  # RegulonDB can return 'nan' or '' gene names
                        bnum = get_gene_bnum(gene_id, gene_synonym_df)
                        _add_links_to_d(bnum, pathway_operon_d, op_ID_set)
    return pathway_operon_d

                    
all_muts_df["pathway links"] = all_muts_df.apply(get_pathway_operon_links_d, axis=1)
all_muts_df.head()

all_muts_df.to_pickle("./data/2_5_df.pkl")
