==> 4_5_assoc_genetic_feat_to_cond.ipynb <==
%matplotlib inline
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd
import seaborn as sns
from IPython.display import display
import os, sys, itertools, csv
module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)
from mutil.alemutdf import get_all_sample_mut_df, get_gene_mut_count_mat, get_multi_exp_max_freq_mut_df, get_mut_type_avg_frac_across_class_df
from mutil.metadata import get_condition_val_dict, get_condition_field_val_set
from mutil.genome import get_K12_pos_from_BOP27, NON_K12_EXP_L
from mutil.params import ASSOC_ALPHA, MULTI_HYP_CORR_METHOD
pd.options.display.max_columns = 100

# starting efforts for refactoring of assoc NBs.
FEAT_ANNOT = "genetic features"
LINK_ANNOT = "genetic feature links"

all_muts_df = pd.read_pickle("./data/4_df.pkl")
display(all_muts_df.shape, all_muts_df.head())

from mutil.metadata import get_all_exp_cond_d

all_exp_cond_d = get_all_exp_cond_d("./data/metadata/")
all_exp_cond_d["TOL_2,3-butanediol"] = all_exp_cond_d.pop("TOL_2 3-butanediol")  # TODO: workaround for bug of missing comma. Fix root bug
all_exp_cond_d

# So that rows will always contain the same amount of columns
all_muts_df = all_muts_df.replace('', "None")
display(all_muts_df.shape, all_muts_df.head())

from mutil.metadata_categories_for_associations import METADATA_CATEGORIES_FOR_ASSOCIATIONS

for condition in METADATA_CATEGORIES_FOR_ASSOCIATIONS:
    all_muts_df[condition] = all_muts_df["exp"].apply(lambda exp: all_exp_cond_d[exp][condition])
    
all_muts_df = all_muts_df.reset_index(drop=True)

# The below is the necessary logic without this filtering.
exp_target_cond_df = all_muts_df.copy()
display(exp_target_cond_df.shape, exp_target_cond_df.head())

unique_condition_set = set()
for cond_cat in METADATA_CATEGORIES_FOR_ASSOCIATIONS:
    unique_condition_set |= set(exp_target_cond_df[cond_cat].unique())
unique_condition_set

from mutil.feature import get_feat_d


feat_cond_df = pd.DataFrame()
for _, r in exp_target_cond_df.iterrows():

    # This is the only annotation level that we should be using mutations from experiments using a strain other than K12 derived.
    if r["exp"] in NON_K12_EXP_L:
        for gen_feat in r[FEAT_ANNOT]:
            for cond_col in METADATA_CATEGORIES_FOR_ASSOCIATIONS:
                if r[cond_col] in unique_condition_set:  # I'm not quite sure why I had to check this in the legacy code.
                    feat_cond_df = feat_cond_df.append(
                        {"feature": gen_feat["name"], "condition": r[cond_col]}, ignore_index=True)
    else:
        for gen_feat_ID, links in r[LINK_ANNOT].items():
            for cond_col in METADATA_CATEGORIES_FOR_ASSOCIATIONS:
                if r[cond_col] in unique_condition_set:  # I'm not quite sure why I had to check this in the legacy code.
                    # appending the genetic feature to the feat_cond_df as many times as it has links to it
                    # the links represent the number of times the genetic feat was mutated.
                    # same approach as that of mutflow feature counts.
                    gen_feat_d = get_feat_d(RegulonDB_ID=gen_feat_ID, json=r[FEAT_ANNOT])
                    df = pd.DataFrame([{"feature": gen_feat_d["name"], "condition": r[cond_col]} for _ in range(0, len(links))])
                    feat_cond_df = feat_cond_df.append(df, ignore_index=True)

display(len(feat_cond_df), feat_cond_df.head())

from statsmodels.stats import multitest
import scipy


def get_contingency_table(count_mat, row_name, col_name):
    row_col_count = count_mat.loc[row_name, col_name]
    row_not_col_sum_count = count_mat.loc[row_name].sum() - row_col_count
    not_row_col_sum_count = count_mat.T.loc[col_name].sum() - row_col_count
    all_mat_sum_count = count_mat.sum(axis=1).sum()
    not_row_not_col_sum_count = all_mat_sum_count - \
        row_col_count - row_not_col_sum_count - not_row_col_sum_count
    contingency_table = [
        [row_col_count, row_not_col_sum_count],
        [not_row_col_sum_count, not_row_not_col_sum_count]
    ]
    return contingency_table


test_df = pd.DataFrame([[1, 2, 3],
                        [4, 5, 6]],
                       index=["y1", "y2"],
                       columns=["x1", "x2", "x3"])
contingency_table = get_contingency_table(test_df, "y2", "x3")
expected_contingency_table = [[6, 9], [3, 3]]
assert(contingency_table == expected_contingency_table)


def get_multiple_hypothesis_correction(pval_df):
    # Dataframe seemed to be designed to iterate by column more easily than row (see df.iteritems())
    # Have to combine all p-value columns because multitest.multipletests only takes 1D array inputs.
    pval_l = []
    for col in list(pval_df.columns.values):
        pval_l += list(pval_df[col])
    corrected_pval_result = multitest.multipletests(
        pvals=pval_l,
        alpha=ASSOC_ALPHA,
        method=MULTI_HYP_CORR_METHOD)

    # The following splits the multitest.multipletests 1D array output into the same shape, though transposed.
    corrected_pval_l = corrected_pval_result[1]
    df_col_len = pval_df.shape[0]
    corrected_pval_mat = [corrected_pval_l[i:i+df_col_len]
                          for i in range(0, len(corrected_pval_l), df_col_len)]

    # building df with new p-values
    corrected_pval_df = pval_df.copy()
    for outer_l_idx in range(0, len(corrected_pval_df.columns)):
        for inner_l_idx in range(0, len(corrected_pval_df)):
            corrected_pval = corrected_pval_mat[outer_l_idx][inner_l_idx]
            corrected_pval_df.iloc[inner_l_idx, outer_l_idx] = corrected_pval
    return corrected_pval_df


# def get_enrich_genetic_target_df(cond_col, mut_df):
def get_enrich_genetic_target_df(feat_cond_df):
    # Creates tables of counts between each mutation and unique condition.
    cross_counts_df = pd.crosstab(
        feat_cond_df["feature"], feat_cond_df["condition"]
    )

    # To reuse the same DF indeces without having to remake it.
    enrich_odds_ratio_df = cross_counts_df.copy()
    enrich_pvals_df = cross_counts_df.copy()

    for cond in cross_counts_df.columns.values:
        for mut_target in cross_counts_df.index:
            contingency_table = get_contingency_table(cross_counts_df, mut_target, cond)
            odds_ratio, p_val = scipy.stats.fisher_exact(contingency_table, alternative="greater")
            enrich_odds_ratio_df.loc[mut_target, cond] = odds_ratio
            enrich_pvals_df.loc[mut_target, cond] = p_val

    enriched_pvals_df = get_multiple_hypothesis_correction(enrich_pvals_df)

    return enrich_odds_ratio_df, enrich_pvals_df

enrich_odds_ratio_df, enrich_pvals_df = get_enrich_genetic_target_df(feat_cond_df)

signif_genomic_feat_cond_json = []
mut_feat_cond_assoc_json = []
for (mut_serial, row) in enrich_pvals_df.iterrows():
    for cond in row.index:
        p_val = enrich_pvals_df.at[mut_serial, cond]
        odds_ratio = enrich_odds_ratio_df.at[mut_serial, cond]
        mut_feat_cond_assoc_json.append({"mutated features": mut_serial, "condition": cond, "odd ratio": odds_ratio, "p value":p_val})
        if odds_ratio > 1 and p_val < ASSOC_ALPHA:
            signif_genomic_feat_cond_json.append({"mutated features": mut_serial, "condition": cond, "odd ratio": odds_ratio, "p value":p_val})

assoc_df = pd.DataFrame()
for d in mut_feat_cond_assoc_json:
    assoc_df = assoc_df.append(d, ignore_index=True)
assoc_df = assoc_df.set_index("mutated features")
assoc_df.to_csv("./data/supp/mut_gen_cond_assocs.csv")

import pickle
f = open("./data/signif_genetic_feat_cond_json.pkl", 'wb')
pickle.dump(signif_genomic_feat_cond_json, f)
f.close()

signf_assoc_feat_cond_d = {d["mutated features"]:set() for d in signif_genomic_feat_cond_json}
for d in signif_genomic_feat_cond_json:
    target = d["mutated features"]
    condition = d["condition"]
    if target in signf_assoc_feat_cond_d.keys():
        signf_assoc_feat_cond_d[target].add(condition)
    else:
        signf_assoc_feat_cond_d[target] = {condition}
# signf_assoc_feat_cond_d

def _get_feat_signf_assoc_cond_set(feat_d, mut_conds):
    feat_signf_assoc_cond_set = set()
    if feat_d["name"] in signf_assoc_feat_cond_d.keys():
        
        # Only includes conditions from exp mut comes from, therefore may be subset of signf assoc cond.
        for c in signf_assoc_feat_cond_d[feat_d["name"]]:
            if c in mut_conds:
                feat_signf_assoc_cond_set.add(c)
    
    return feat_signf_assoc_cond_set


def get_feat_signf_assoc_cond_set_json(mut_df_row, feat_annot_name):
    feat_json = []
    for d in mut_df_row[feat_annot_name]:
        feat_d = d.copy()
        mut_conds = list(mut_df_row[METADATA_CATEGORIES_FOR_ASSOCIATIONS])
        feat_d["significantly associated conditions"] = _get_feat_signf_assoc_cond_set(d, mut_conds)
        feat_json.append(feat_d)
    return feat_json


exp_target_cond_df[FEAT_ANNOT] = exp_target_cond_df.apply(lambda r: get_feat_signf_assoc_cond_set_json(r, FEAT_ANNOT), axis=1)

exp_target_cond_df.to_pickle("./data/4_5_df.pkl")
exp_target_cond_df.shape
